% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical-elements.R
\name{make_bandscale}
\alias{make_bandscale}
\title{Make a bandscale}
\usage{
make_bandscale(range, n, pad_in = 0, pad_out = 0, reverse = F)
}
\arguments{
\item{range}{<num> length of the range of the data}

\item{n}{<num> number of steps in the range}

\item{pad_in}{<num> fraction of step size to pad in between bands, defaults
to 0.}

\item{pad_out}{<num> fraction of step size to pad outside of first and last
band, defaults to 0.}

\item{reverse}{<lgl> if the values of the scale should be reversed. Useful for y-down vs . y-up}
}
\value{
A \code{\link[=list]{list()}} with elements: \link{range}, \link{n}, \link{step}, \link{bandwidth}, \link{pad_in}
, \link{pad_out}, \link{first}, \link{values}.
}
\description{
Make a bandscale
}
\examples{
    make_bandscale(127.76, 12, 0.02, 0.17) # column bands for 96 well plate
    make_bandscale(87.45, 8, 0.02, 0.245) # row bands for 96 well plate
}
