% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical-elements.R
\name{make_footprint}
\alias{make_footprint}
\title{footprint of microtitre plate from width and height}
\usage{
make_footprint(w, h, r = 3)
}
\arguments{
\item{w}{<num> width of plate}

\item{h}{<num> height of plate}

\item{r}{<num> radius of circle used to round corners. currently not used.}
}
\value{
A \code{\link[=tibble]{tibble()}} of length 1. Each row is a rect with variables: \link{x},
\link{y}, \link{w}, \link{h}, \link{rx}, \link{ry}, \link{xmin}, \link{xmax}, \link{ymin}, \link{ymax}.
}
\description{
footprint of microtitre plate from width and height
}
\examples{
    make_footprint(127.76, 87.45) # SBS plate footprint.
}
