% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical-elements.R
\name{make_notched_border}
\alias{make_notched_border}
\title{notched well border of microtitre plate from width and height}
\usage{
make_notched_border(w, h)
}
\arguments{
\item{w}{<num> width of plate}

\item{h}{<num> height of plate}
}
\value{
A \code{\link[=tibble]{tibble()}} of length 7. Each row is a point with coordinates: \link{x},
\link{y}.
}
\description{
notched well border of microtitre plate from width and height
}
\examples{
    make_notched_border(127.76, 87.45) # SBS plate footprint.
}
