% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical-elements.R
\name{make_row_labels}
\alias{make_row_labels}
\alias{make_col_labels}
\title{make row labels for a microtitre plate from a set of scales}
\usage{
make_row_labels(xs, ys, nudge_x = 0)

make_col_labels(xs, ys, nudge_y = 0)
}
\arguments{
\item{xs}{<list> bandwidth scale trained on plate height and n cols}

\item{ys}{<list> bandwidth scale trained on plate width and n rows}

\item{nudge_x}{<num>  amount to nudge the labels by, defaults to 0.}

\item{nudge_y}{<num>  amount to nudge the labels by, defaults to 0.}
}
\value{
A \code{\link[=tibble]{tibble()}} of length nrows. One row per label with \link{label}
positioned at point \link{x}, \link{y}.
}
\description{
make row labels for a microtitre plate from a set of scales
}
