% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical-elements.R
\name{make_wells}
\alias{make_wells}
\title{Make rectangular wells of a microtire plate from trained scales.}
\usage{
make_wells(xs, ys)
}
\arguments{
\item{xs}{<list> bandwidth scale trained on plate height and n cols}

\item{ys}{<list> bandwidth scale trained on plate width and n rows}
}
\value{
A \code{\link[=tibble]{tibble()}} with variables: \link{xmin}, \link{xmax}, \link{ymin}, \link{ymax}.
}
\description{
Make rectangular wells of a microtire plate from trained scales.
}
