% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp-view.R
\name{mtp_view}
\alias{mtp_view}
\title{View data laid out in microtitre plate.
Tooltip enabled by d3 and r2d3.}
\usage{
mtp_view(tbl, fillVar = NULL, fillOpacityVar = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{tbl}{table of data to use which must have well col in 'A01' style.}

\item{fillVar}{name of variable in tbl to use to map the fill of each well to}

\item{fillOpacityVar}{name of variable in tbl to use to map the fill opacity to}

\item{width}{width of the resulting mtp}

\item{height}{height of the resulting mtp}
}
\description{
View data laid out in microtitre plate.
Tooltip enabled by d3 and r2d3.
}
